# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CSR Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ORGANIZATIONS
...
  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List proposals
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: province
          in: query
          schema:
            type: ["string", "null"]
        - name: organizationId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Proposal list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProposalListResponse"
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Create proposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProposalBody"
      responses:
        "201":
          description: Created proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}:
    get:
      operationId: getProposal
      tags: [proposals]
      summary: Get proposal detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProposalDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProposal
      tags: [proposals]
      summary: Update proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProposalBody"
      responses:
        "200":
          description: Updated proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}/status:
    patch:
      operationId: updateProposalStatus
      tags: [proposals]
      summary: Update proposal status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProposalStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}/ai-score:
    get:
      operationId: getProposalAiScore
      tags: [proposals]
      summary: Get AI score and recommendations for proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI score result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiScoreResponse"

  # PROJECTS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        emailVerified:
          type: boolean
        organizationId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - isActive
        - emailVerified
        - createdAt
        - updatedAt
    UpdateUserBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit

    # ORGANIZATIONS
...
    Proposal:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        category:
          type: string
        status:
          type: string
        organizationId:
          type: integer
        organizationName:
          type: string
        province:
          type: string
        city:
          type: ["string", "null"]
        targetBeneficiaries:
          type: integer
        budgetTotal:
          type: integer
        fundedAmount:
          type: integer
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        sdgGoals:
          type: array
          items:
            type: integer
        tags:
          type: array
          items:
            type: string
        aiScore:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - summary
        - category
        - status
        - organizationId
        - organizationName
        - province
        - targetBeneficiaries
        - budgetTotal
        - fundedAmount
        - sdgGoals
        - tags
        - createdAt
        - updatedAt
    ProposalDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
        organizationId:
          type: integer
        organizationName:
          type: string
        province:
          type: string
        city:
          type: ["string", "null"]
        targetBeneficiaries:
          type: integer
        budgetTotal:
          type: integer
        fundedAmount:
          type: integer
        budgetBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/BudgetItem"
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        sdgGoals:
          type: array
          items:
            type: integer
        tags:
          type: array
          items:
            type: string
        aiScore:
          type: ["integer", "null"]
        aiSummary:
          type: ["string", "null"]
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/StatusHistoryItem"
        cofunding:
          type: array
          items:
            $ref: "#/components/schemas/CofundingCommitment"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - summary
        - description
        - category
        - status
        - organizationId
        - organizationName
        - province
        - targetBeneficiaries
        - budgetTotal
        - fundedAmount
        - budgetBreakdown
        - sdgGoals
        - tags
        - statusHistory
        - cofunding
        - createdAt
        - updatedAt
    BudgetItem:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        amount:
          type: integer
        category:
          type: string
      required:
        - id
        - description
        - amount
        - category
    StatusHistoryItem:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        notes:
          type: ["string", "null"]
        changedBy:
          type: string
        createdAt:
          type: string
      required:
        - id
        - status
        - changedBy
        - createdAt
    CreateProposalBody:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        description:
          type: string
        category:
          type: string
        province:
          type: string
        city:
          type: ["string", "null"]
        targetBeneficiaries:
          type: integer
        budgetTotal:
          type: integer
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        sdgGoals:
          type: array
          items:
            type: integer
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - summary
        - description
        - category
        - province
        - targetBeneficiaries
        - budgetTotal
    UpdateProposalBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        targetBeneficiaries:
          type: ["integer", "null"]
        budgetTotal:
          type: ["integer", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        sdgGoals:
          type: array
          items:
            type: integer
        tags:
          type: array
          items:
            type: string
    UpdateProposalStatusBody:
      type: object
      properties:
        status:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - status
    ProposalListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Proposal"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    AiScoreResponse:
      type: object
      properties:
        score:
          type: integer
        completeness:
          type: integer
        relevance:
          type: integer
        impact:
          type: integer
        riskLevel:
          type: string
        summary:
          type: string
        recommendations:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        sdgAlignment:
          type: array
          items:
            type: integer
      required:
        - score
        - completeness
        - relevance
        - impact
        - riskLevel
        - summary
        - recommendations
        - tags
        - sdgAlignment

    # PROJECTS
...
    CofundingCommitment:
      type: object
      properties:
        id:
          type: integer
        proposalId:
          type: integer
        organizationId:
          type: integer
        organizationName:
          type: string
        amount:
          type: integer
        percentageShare:
          type: number
        status:
          type: string
        notes:
          type: ["string", "null"]
        committedAt:
          type: string
        confirmedAt:
          type: ["string", "null"]
      required:
        - id
        - proposalId
        - organizationId
        - organizationName
        - amount
        - percentageShare
        - status
        - committedAt
...