# lib/api-zod/src/generated/types/budgetItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CSR Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BudgetItem {
  id: number;
  description: string;
  amount: number;
  category: string;
}